First inscribed circle

+++ stepper/function.h
@@ -1,6 +1,7 @@
 #pragma once
 
-#include "Arduino.h"
+#include <Arduino.h>
+
 #include "stepper_lib.h"
 
 typedef struct {
@@ -12,6 +13,6 @@ typedef struct {
 #define STEP_STOP (Step2D) { true }
 #define STEP(x, y) (Step2D) { false, x, y }
 
-typedef Step2D (*Function2D)(unsigned t, unsigned x, unsigned y);
+typedef Step2D (*Function2D)(long t, unsigned x, unsigned y);
 
 void apply_function(Axes2D * axes, Function2D function);
